botservice/requirements.txt
openai
langchain
langchain-openai
langchain-community
langchain-experimental
langgraph
botbuilder-integration-aiohttp
tiktoken
pypdf
sqlalchemy
pyodbc
tabulate
azure-cosmos
streamlit
audio-recorder-streamlit
python-dotenv
azure-ai-formrecognizer
azure-storage-blob
beautifulsoup4
pydantic
azure-cognitiveservices-speech
gunicorn
